MyHearingBuddy/sign-language-detector/requirements.txt
opencv-python>=4.7.0
mediapipe>=0.10.13
scikit-learn>=1.2.0
openai>=0.28.1
pyttsx3>=2.90
flask>=2.3.0
flask-cors>=4.0.0
pillow>=9.0.0
numpy>=1.21.0
